feat: Implement github fetcher
